refactor: quiz screen ui
