+++ features/quiz/ui/src/main/java/ru/popkov/russeliterature/features/quiz/ui/QuizScreen.kt
@@ -2,9 +2,11 @@ package ru.popkov.russeliterature.features.quiz.ui
 
 import androidx.compose.foundation.Image
 import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.fillMaxWidth
@@ -20,11 +22,9 @@ import androidx.compose.material.icons.filled.Close
 import androidx.compose.material3.Button
 import androidx.compose.material3.ButtonDefaults
 import androidx.compose.material3.Icon
-import androidx.compose.material3.IconButton
 import androidx.compose.material3.SnackbarHostState
 import androidx.compose.material3.Text
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.Immutable
 import androidx.compose.runtime.LaunchedEffect
 import androidx.compose.runtime.collectAsState
 import androidx.compose.runtime.getValue
@@ -35,6 +35,7 @@ import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.style.TextAlign
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
@@ -67,52 +68,25 @@ fun QuizScreen(
     )
 }
 
-@Immutable
-data class QuizAnswer(
-    val id: Int = 0,
-    val answer: String = "",
-    val isCorrect: Boolean = false,
-)
-
-@Immutable
-data class QuizItem(
-    val id: Int = 0,
-    val quizQuestion: String = "В давние времена Нижний был центром соляной промышленности. В 1864 году соляные склады оказались пустыми.",
-    val quizAnswer: String = "В то время действовала монополия на продажу соли, а нечистые на руку купцы потихоньку распродавали её и списывали убытки на паводки. Склады стояли у реки и, действительно, могли подновляться. Афера удачно складывалась до того, как Александр II отменил монополию и приказал срочно продать соль. Паводка в том году не было, так что свалить убытки было не на что. Участники аферы осудили по всей строгости закона. По городской легенде, причастному к скандалу купцу.",
-    val answers: List<QuizAnswer> = listOf(
-        QuizAnswer(
-            id = 0,
-            answer = "Признались в краже соли",
-            isCorrect = true,
-        ),
-        QuizAnswer(
-            id = 1,
-            answer = "Паводок уничтожил соль",
-            isCorrect = false,
-        ),
-        QuizAnswer(
-            id = 2,
-            answer = "Отгрузили соль в долг до паводка",
-            isCorrect = false,
-        ),
-    ),
-)
-
 @Composable
 fun Header(
     modifier: Modifier = Modifier,
     onCloseClick: (QuizViewAction) -> Unit = {},
 ) {
-    IconButton(
+    Box(
         modifier = modifier
+            .size(30.dp)
             .clip(shape = RoundedCornerShape(size = 8.dp))
-            .background(color = Colors.ButtonCloseColor),
-        onClick = { onCloseClick(QuizViewAction.OnCloseClick) },
+            .background(color = Colors.ButtonCloseColor)
+            .clickable { onCloseClick(QuizViewAction.OnCloseClick) },
+        contentAlignment = Alignment.Center,
     ) {
         Icon(
+            modifier = Modifier
+                .size(24.dp),
             imageVector = Icons.Filled.Close,
+            contentDescription = "Close",
             tint = Color.White,
-            contentDescription = null,
         )
     }
 }
@@ -146,6 +120,8 @@ fun Question(
             ) {
                 if (icon != null) {
                     Icon(
+                        modifier = Modifier
+                            .size(size = 18.dp),
                         imageVector = icon,
                         tint = Color.White,
                         contentDescription = null,
@@ -159,7 +135,7 @@ fun Question(
             }
             Text(
                 modifier = Modifier
-                    .padding(all = 10.dp),
+                    .padding(horizontal = 14.dp, vertical = 4.dp),
                 text = quizAnswer.answer,
                 style = FormularRegular16,
             )
@@ -202,51 +178,63 @@ internal fun Quiz(
             }
 
             else -> {
-                Image(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(top = 26.dp)
-                        .height(height = 240.dp)
-                        .clip(shape = RoundedCornerShape(size = 16.dp)),
-                    painter = painterResource(id = ru.popkov.android.core.feature.ui.R.drawable.ic_article),
-                    contentDescription = null,
-                    contentScale = ContentScale.Crop,
-                )
-                Text(
-                    modifier = Modifier
-                        .padding(top = 20.dp, bottom = 6.dp)
-                        .align(Alignment.Start),
-                    text = "ПРАВИЛЬНЫЙ ОТВЕТ:",
-                    style = FormularRegular12,
-                    color = Color.White.copy(alpha = 0.7f),
-                )
-                Question(
-                    quizAnswer = state.item.answers.find { it.isCorrect }!!,
+                Result(
+                    state = state,
                     onAnswerClick = onAnswerClick,
-                    icon = Icons.Filled.Check,
-                )
-                Text(
-                    modifier = Modifier
-                        .padding(top = 20.dp, bottom = 6.dp)
-                        .align(Alignment.Start),
-                    text = "ВАШ ОТВЕТ:",
-                    style = FormularRegular12,
-                    color = Color.White.copy(alpha = 0.7f),
-                )
-                Question(
-                    quizAnswer = state.item.answers.find { !it.isCorrect }!!,
-                    onAnswerClick = onAnswerClick,
-                    icon = Icons.Filled.Close,
-                )
-                Text(
-                    modifier = Modifier
-                        .padding(vertical = 16.dp),
-                    text = state.item.quizAnswer,
-                    style = FormularRegular14,
                 )
             }
         }
     }
+}
+
+@Composable
+internal fun ColumnScope.Result(
+    modifier: Modifier = Modifier,
+    state: QuizState,
+    onAnswerClick: (QuizViewAction) -> Unit = {},
+) {
+    Image(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(top = 50.dp)
+            .height(height = 240.dp)
+            .clip(shape = RoundedCornerShape(size = 16.dp)),
+        painter = painterResource(id = ru.popkov.android.core.feature.ui.R.drawable.ic_article),
+        contentDescription = null,
+        contentScale = ContentScale.Crop,
+    )
+    Text(
+        modifier = Modifier
+            .padding(top = 20.dp, bottom = 6.dp)
+            .align(Alignment.Start),
+        text = stringResource(id = R.string.quiz_right),
+        style = FormularRegular12,
+        color = Color.White.copy(alpha = 0.7f),
+    )
+    Question(
+        quizAnswer = state.item.answers.find { it.isCorrect }!!,
+        onAnswerClick = onAnswerClick,
+        icon = Icons.Filled.Check,
+    )
+    Text(
+        modifier = Modifier
+            .padding(top = 10.dp, bottom = 6.dp)
+            .align(Alignment.Start),
+        text = stringResource(id = R.string.quiz_answer),
+        style = FormularRegular12,
+        color = Color.White.copy(alpha = 0.7f),
+    )
+    Question(
+        quizAnswer = state.item.answers.find { !it.isCorrect }!!,
+        onAnswerClick = onAnswerClick,
+        icon = Icons.Filled.Close,
+    )
+    Text(
+        modifier = Modifier
+            .padding(vertical = 16.dp),
+        text = state.item.quizAnswer,
+        style = FormularRegular14,
+    )
 }
 
 @Preview(showBackground = true)
